separa row do list
