--- PictureDay/Views/FavoriteRowView.swift
@@ -1,8 +1,0 @@
-//
-//  FavoriteRowView.swift
-//  PictureDay
-//
-//  Created by Kleyson Tavares on 16/09/25.
-//
-
-import Foundation
