--- PictureDay/Views/FavoritesView.swift
@@ -1,189 +1,0 @@
-//
-//  FavoritesView.swift
-//  PictureDay
-//
-//  Created by Kleyson Tavares on 16/09/25.
-//
-
-import SwiftUI
-
-struct FavoritesView: View {
-    @StateObject private var viewModel: FavoritesViewModel
-    @State private var selectedAPOD: APODModel?
-    
-    init(favoritesService: FavoritesServiceProtocol) {
-        self._viewModel = StateObject(wrappedValue: FavoritesViewModel(favoritesService: favoritesService))
-    }
-    
-    var body: some View {
-        NavigationView {
-            ZStack {
-                Color.black.ignoresSafeArea()
-                
-                if viewModel.isLoading {
-                    ProgressView("Carregando favoritos...")
-                        .foregroundColor(.white)
-                        .font(.headline)
-                } else if viewModel.favorites.isEmpty {
-                    VStack(spacing: 20) {
-                        Image(systemName: "heart")
-                            .font(.system(size: 50))
-                            .foregroundColor(.gray)
-                        
-                        Text("Nenhum favorito ainda")
-                            .font(.title2)
-                            .fontWeight(.bold)
-                            .foregroundColor(.white)
-                        
-                        Text("Toque no coração nas fotos para adicionar aos favoritos")
-                            .font(.body)
-                            .foregroundColor(.gray)
-                            .multilineTextAlignment(.center)
-                            .padding(.horizontal)
-                    }
-                } else {
-                    List {
-                        ForEach(viewModel.favorites) { apod in
-                            FavoriteRowView(
-                                apod: apod,
-                                onTap: {
-                                    selectedAPOD = apod
-                                },
-                                onRemove: {
-                                    viewModel.removeFromFavorites(apod)
-                                }
-                            )
-                            .listRowBackground(Color.clear)
-                            .listRowSeparator(.hidden)
-                        }
-                    }
-                    .listStyle(PlainListStyle())
-                    .refreshable {
-                        viewModel.fetchFavorites()
-                    }
-                }
-                
-                if let errorMessage = viewModel.errorMessage {
-                    VStack {
-                        Spacer()
-                        Text(errorMessage)
-                            .foregroundColor(.red)
-                            .padding()
-                            .background(Color.black.opacity(0.8))
-                            .cornerRadius(8)
-                            .padding()
-                    }
-                }
-            }
-            .navigationTitle("Favoritos")
-            .navigationBarTitleDisplayMode(.large)
-            .onAppear {
-                viewModel.fetchFavorites()
-            }
-            .sheet(item: $selectedAPOD) { apod in
-                PictureDetailView(apod: apod, favoritesService: MockFavoritesService())
-            }
-        }
-    }
-}
-
-// MARK: - Favorite Row View
-struct FavoriteRowView: View {
-    let apod: APODModel
-    let onTap: () -> Void
-    let onRemove: () -> Void
-    
-    @State private var thumbnailImage: UIImage?
-    @State private var isLoadingImage = true
-    
-    var body: some View {
-        Button(action: onTap) {
-            HStack(spacing: 12) {
-                // Thumbnail da imagem
-                Group {
-                    if let image = thumbnailImage {
-                        Image(uiImage: image)
-                            .resizable()
-                            .aspectRatio(contentMode: .fill)
-                    } else if isLoadingImage {
-                        ProgressView()
-                            .frame(width: 80, height: 80)
-                    } else {
-                        Rectangle()
-                            .fill(Color.gray.opacity(0.3))
-                            .overlay(
-                                Image(systemName: "photo")
-                                    .foregroundColor(.gray)
-                            )
-                    }
-                }
-                .frame(width: 80, height: 80)
-                .cornerRadius(8)
-                .clipped()
-                
-                VStack(alignment: .leading, spacing: 4) {
-                    Text(apod.title)
-                        .font(.headline)
-                        .foregroundColor(.white)
-                        .lineLimit(2)
-                        .multilineTextAlignment(.leading)
-                    
-                    Text(formatDate(apod.date))
-                        .font(.caption)
-                        .foregroundColor(.gray)
-                    
-                    Text(apod.explanation)
-                        .font(.caption)
-                        .foregroundColor(.white.opacity(0.8))
-                        .lineLimit(3)
-                        .multilineTextAlignment(.leading)
-                }
-                
-                Spacer()
-                
-                Button(action: {
-                    onRemove()
-                }) {
-                    Image(systemName: "heart.fill")
-                        .foregroundColor(.red)
-                        .font(.title3)
-                }
-                .buttonStyle(PlainButtonStyle())
-            }
-            .padding(.vertical, 8)
-        }
-        .buttonStyle(PlainButtonStyle())
-        .onAppear {
-            loadThumbnail()
-        }
-    }
-    
-    private func loadThumbnail() {
-        guard let imageURL = URL(string: apod.url) else { return }
-        
-        URLSession.shared.dataTask(with: imageURL) { data, response, error in
-            DispatchQueue.main.async {
-                if let data = data, let uiImage = UIImage(data: data) {
-                    self.thumbnailImage = uiImage
-                }
-                self.isLoadingImage = false
-            }
-        }.resume()
-    }
-    
-    private func formatDate(_ dateString: String) -> String {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "yyyy-MM-dd"
-        
-        guard let date = formatter.date(from: dateString) else {
-            return dateString
-        }
-        
-        formatter.dateFormat = "dd/MM/yyyy"
-        return formatter.string(from: date)
-    }
-}
-
-#Preview {
-    FavoritesView(favoritesService: MockFavoritesService())
-}
