+++ PictureDay/Views/ListRowView.swift
@@ -1,0 +1,105 @@
+//
+//  ListRowView.swift
+//  PictureDay
+//
+//  Created by Kleyson Tavares on 16/09/25.
+//
+
+import SwiftUI
+
+struct ListRowView: View {
+    let apod: APODModel
+    let isFavorite: Bool
+    let onFavoriteToggle: () -> Void
+    let onTap: () -> Void
+    
+    @State private var thumbnailImage: UIImage?
+    @State private var isLoadingImage = true
+    
+    var body: some View {
+        Button(action: onTap) {
+            HStack(spacing: 12) {
+                // Thumbnail da imagem
+                Group {
+                    if let image = thumbnailImage {
+                        Image(uiImage: image)
+                            .resizable()
+                            .aspectRatio(contentMode: .fill)
+                    } else if isLoadingImage {
+                        ProgressView()
+                            .frame(width: 80, height: 80)
+                    } else {
+                        Rectangle()
+                            .fill(Color.gray.opacity(0.3))
+                            .overlay(
+                                Image(systemName: "photo")
+                                    .foregroundColor(.gray)
+                            )
+                    }
+                }
+                .frame(width: 80, height: 80)
+                .cornerRadius(8)
+                .clipped()
+                
+                VStack(alignment: .leading, spacing: 4) {
+                    Text(apod.title)
+                        .font(.headline)
+                        .foregroundColor(.white)
+                        .lineLimit(2)
+                        .multilineTextAlignment(.leading)
+                    
+                    Text(formatDate(apod.date))
+                        .font(.caption)
+                        .foregroundColor(.gray)
+                    
+                    Text(apod.explanation)
+                        .font(.caption)
+                        .foregroundColor(.white.opacity(0.8))
+                        .lineLimit(3)
+                        .multilineTextAlignment(.leading)
+                }
+                
+                Spacer()
+                
+                Button(action: {
+                    onFavoriteToggle()
+                }) {
+                    Image(systemName: isFavorite ? "heart.fill" : "heart")
+                        .foregroundColor(isFavorite ? .red : .white)
+                        .font(.title3)
+                }
+                .buttonStyle(PlainButtonStyle())
+            }
+            .padding(.vertical, 8)
+        }
+        .buttonStyle(PlainButtonStyle())
+        .onAppear {
+            loadThumbnail()
+        }
+    }
+    
+    private func loadThumbnail() {
+        guard let imageURL = URL(string: apod.url) else { return }
+        
+        URLSession.shared.dataTask(with: imageURL) { data, response, error in
+            DispatchQueue.main.async {
+                if let data = data, let uiImage = UIImage(data: data) {
+                    self.thumbnailImage = uiImage
+                }
+                self.isLoadingImage = false
+            }
+        }.resume()
+    }
+    
+    private func formatDate(_ dateString: String) -> String {
+        let formatter = DateFormatter()
+        formatter.dateFormat = "yyyy-MM-dd"
+        
+        guard let date = formatter.date(from: dateString) else {
+            return dateString
+        }
+        
+        formatter.dateFormat = "dd/MM/yyyy"
+        return formatter.string(from: date)
+    }
+}
